+++ src/ace_segment/testing/TestableLedModule.h
@@ -1,7 +1,7 @@
 /*
 MIT License
 
-Copyright (c) 2018 Brian T. Park
+Copyright (c) 2021 Brian T. Park
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -22,30 +22,26 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-#ifndef ACE_SEGMENT_TESTABLE_LED_DISPLAY_H
-#define ACE_SEGMENT_TESTABLE_LED_DISPLAY_H
+#ifndef ACE_SEGMENT_TESTABLE_LED_MODULE_H
+#define ACE_SEGMENT_TESTABLE_LED_MODULE_H
 
-#include "../LedDisplay.h"
+#include "../LedModule.h"
 
 namespace ace_segment {
 namespace testing {
 
 /**
- * An implementation of LedDisplay for unit testing purposes.
- * Implements most of the simple features of ScanningDisplay.
+ * An implementation of LedModule for unit testing purposes.
  *
  * @tparam DIGITS number of digits supported by this class
  */
 template <uint8_t DIGITS>
-class TestableLedDisplay : public LedDisplay {
+class TestableLedModule : public LedModule {
   public:
-    TestableLedDisplay() : LedDisplay(DIGITS) {}
+    TestableLedModule() : LedModule(DIGITS) {}
 
     void setBrightness(uint8_t /*brightness*/) override {}
 
-    uint8_t* getPatterns() { return mPatterns; }
-
-  protected:
     void setPatternAt(uint8_t pos, uint8_t pattern) override {
       mPatterns[pos] = pattern;
     }
@@ -53,6 +49,8 @@ class TestableLedDisplay : public LedDisplay {
     uint8_t getPatternAt(uint8_t pos) override {
       return mPatterns[pos];
     }
+
+    uint8_t* getPatterns() { return mPatterns; }
 
   private:
     uint8_t mPatterns[DIGITS + 1]; // + 1 to test overflow